add - worked, to fix displaying cart

+++ product-service/src/main/java/com/shop/productservice/model/dto/ProductsListInfo.java
@@ -3,6 +3,7 @@ package com.shop.productservice.model.dto;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -12,7 +13,7 @@ public class ProductsListInfo {
     private List<ProductDto> products;
     private long totalElements;
     private List<String> categories;
-    private double maxPrice;
+    private BigDecimal maxPrice;
 
     public ProductsListInfo(List<ProductDto> products, long totalElements) {
         this.products = products;
